#36 fix: 유저 아톰 생성
