+++ src/recoil/groupState.ts
@@ -1,6 +1,6 @@
-import {atom} from 'recoil';
+import { atom } from 'recoil';
 
 export const groupState = atom({
-  key: "groupState",
-  default: 0
-})
+  key: 'groupState',
+  default: 0,
+});
